requirements.txt
requests == 2.32.3
selenium == 4.28.1
websockets == 14.2
python-socketio==4.6.1 # 중요!
pymysql==1.1.1
pandas==2.2.3
aiohttp==3.11.12
fastapi==0.115.8
pydantic==2.10.6
uvicorn==0.34.0